# spec/api/openapi.yaml
openapi: "3.1.0"
info:
  title: Tracium API
  version: 0.1.0
  description: >
    REST API for querying traces and spans collected by the Tracium
    OpenTelemetry-native LLM observability platform. All timestamps are
    Unix epoch milliseconds. All monetary values are USD.
servers:
  - url: "http://localhost:8090"
    description: Local development server
security:
  - bearerAuth: []
paths:
  /v1/traces:
    get:
      operationId: listTraces
      summary: List traces
      description: >
        Returns a paginated list of traces for the authenticated tenant.
        Results are ordered by start_time_ms descending.
      tags:
        - Traces
      parameters:
        - name: tenant_id
          in: query
          description: Filter by tenant ID (admin use — regular callers are scoped to their own tenant).
          schema:
            type: string
        - name: model
          in: query
          description: Filter traces that contain at least one span using this model (normalized model id).
          schema:
            type: string
        - name: agent
          in: query
          description: Filter to traces belonging to a single agent (the trace root name) — powers an agent's recent runs.
          schema:
            type: string
        - name: has_error
          in: query
          description: Filter to traces that have at least one error span.
          schema:
            type: boolean
        - name: start_after
          in: query
          description: Return traces whose start_time_ms is after this ISO 8601 date-time.
          schema:
            type: string
            format: date-time
        - name: start_before
          in: query
          description: Return traces whose start_time_ms is before this ISO 8601 date-time.
          schema:
            type: string
            format: date-time
        - name: page
          in: query
          description: Page number (1-based).
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: page_size
          in: query
          description: Number of results per page.
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        "200":
          description: Paginated list of traces.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedTraceResponse"
        "401":
          description: Missing or invalid bearer token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/traces/{id}:
    get:
      operationId: getTrace
      summary: Get a trace by ID
      description: Returns a single trace including its full span list.
      tags:
        - Traces
      parameters:
        - name: id
          in: path
          required: true
          description: The trace_id of the trace to retrieve.
          schema:
            type: string
      responses:
        "200":
          description: The requested trace with all spans.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TraceDetail"
        "404":
          description: Trace not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid bearer token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: JWT or opaque bearer token issued by the Tracium auth service.
...
  schemas:
    Span:
      type: object
      required:
        - trace_id
        - span_id
        - name
      properties:
        trace_id:
          type: string
          description: Globally unique identifier for the trace this span belongs to.
        span_id:
          type: string
          description: Globally unique identifier for this span.
        parent_span_id:
          type: string
          description: span_id of the parent span, if this is a child span.
        name:
          type: string
          description: Human-readable name for this span.
        start_time_ms:
          type: integer
          description: Unix epoch milliseconds when the span started.
        end_time_ms:
          type: integer
          description: Unix epoch milliseconds when the span ended.
        duration_ms:
          type: integer
          description: Duration of the span in milliseconds.
        model:
          type: string
          description: The model name as reported in gen_ai.request.model.
        finish_reason:
          type: string
          description: Why the model stopped generating.
        input_tokens:
          type: integer
          minimum: 0
          description: Number of input tokens consumed.
        output_tokens:
          type: integer
          minimum: 0
          description: Number of output tokens generated.
        cost_usd:
          type: number
          minimum: 0
          description: Computed cost in USD for this span.
        tenant_id:
          type: string
          description: Tenant identifier.
        model_normalized:
          type: string
          description: Canonical model id after alias resolution.
        schema_version:
          type: integer
          minimum: 1
          description: Schema version of this span record.
        error_type:
          type: string
          description: "Added in schema v2. Short error class or type."
        error_message:
          type: string
          description: "Added in schema v2. Human-readable error message if the span failed."
        kind:
          type: string
          enum: ["agent", "llm", "tool", "chain", "retriever", "embedding"]
          description: >
            Normalized span role, derived by the collector from instrumentation
            attributes (OpenInference openinference.span.kind, Traceloop
            traceloop.span.kind, or the OTel GenAI gen_ai.operation.name).
            Optional: omitted when the collector could not classify the span,
            in which case consumers infer the role from the span's context.
    Trace:
      type: object
      required:
        - trace_id
      properties:
        trace_id:
          type: string
          description: Globally unique identifier for this trace.
        name:
          type: string
          description: Human-readable name for the trace.
        start_time_ms:
          type: integer
          description: Unix epoch milliseconds when the trace started.
        end_time_ms:
          type: integer
          description: Unix epoch milliseconds when the trace ended.
        duration_ms:
          type: integer
          description: Total duration in milliseconds.
        tenant_id:
          type: string
          description: Tenant identifier.
        span_count:
          type: integer
          minimum: 0
          description: Total number of spans in this trace.
        has_error:
          type: boolean
          description: True if any span in this trace recorded an error.
        total_cost_usd:
          type: number
          minimum: 0
          description: Sum of cost_usd across all spans in this trace.
        schema_version:
          type: integer
          minimum: 1
          description: Schema version of this trace record.
    TraceDetail:
      allOf:
        - $ref: "#/components/schemas/Trace"
        - type: object
          required:
            - spans
          properties:
            spans:
              type: array
              description: All spans belonging to this trace, ordered by start_time_ms ascending.
              items:
                $ref: "#/components/schemas/Span"
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: "Machine-readable error code (e.g. NOT_FOUND, UNAUTHORIZED, INVALID_REQUEST)."
        message:
          type: string
          description: Human-readable description of the error.
    PaginatedTraceResponse:
      type: object
      required:
        - items
        - total
        - page
        - page_size
        - has_more
      properties:
        items:
          type: array
          description: The traces on this page.
          items:
            $ref: "#/components/schemas/Trace"
        total:
          type: integer
          minimum: 0
          description: Total number of matching traces across all pages.
        page:
          type: integer
          minimum: 1
          description: Current page number (1-based).
        page_size:
          type: integer
          minimum: 1
          description: Number of items per page.
        has_more:
          type: boolean
          description: True if there are more pages after this one.
...